🐞 fix(codex): 修复供应商 Cloudflare 403 拦截
功能修改
- Codex 上游请求改用通用 curl_cffi 传输，兼容受 Cloudflare 防护的供应商。
- Claude 保留原有客户端名称并复用通用实现，补充 Codex 与传输层回归测试。

影响范围
- 仅调整 Codex 出站传输；重试次数、退避、熔断和供应商切换逻辑保持不变。

验证结果
- Python 472 项、全部 JavaScript 测试、语法检查、EXE 冒烟和真实 ai.hybgzs.com HTTP 200 流式请求验证通过。
