+++ local_proxy/transports/claude.py
@@ -1,133 +1,7 @@
-from __future__ import annotations
+"""Backward-compatible Claude names for the shared curl transport."""
 
-import asyncio
-import codecs
-from collections.abc import AsyncIterator, Mapping, Sequence
-from contextlib import suppress
-from dataclasses import dataclass, field
-from typing import Any
+from .curl import CurlClient, CurlRequest, CurlResponse
 
-import httpx
-from curl_cffi.requests import AsyncSession
-from curl_cffi.requests.exceptions import RequestException
-
-
-@dataclass(frozen=True)
-class ClaudeCurlRequest:
-    method: str
-    url: str
-    params: Sequence[tuple[str, str]] = ()
-    headers: Mapping[str, str] = field(default_factory=dict)
-    content: bytes = b""
-
-
-class ClaudeCurlResponse:
-    def __init__(self, response: Any) -> None:
-        self._response = response
-        self.status_code = int(response.status_code)
-        self.reason_phrase = str(getattr(response, "reason", ""))
-        encoding = str(getattr(response, "encoding", "utf-8") or "utf-8")
-        try:
-            codecs.lookup(encoding)
-        except LookupError:
-            encoding = "utf-8"
-        self.encoding = encoding
-        self.headers = httpx.Headers(
-            (key, value)
-            for key, value in response.headers.items()
-            if key.casefold() not in {"content-encoding", "content-length"}
-        )
-        self.content = b""
-        self.is_stream_consumed = False
-        self._closed = False
-
-    async def aiter_raw(self) -> AsyncIterator[bytes]:
-        try:
-            async for chunk in self._response.aiter_content():
-                if chunk:
-                    yield bytes(chunk)
-        except RequestException as exc:
-            raise httpx.ReadError(str(exc)) from exc
-        finally:
-            self.is_stream_consumed = True
-
-    async def aiter_bytes(self) -> AsyncIterator[bytes]:
-        async for chunk in self.aiter_raw():
-            yield chunk
-
-    async def aclose(self) -> None:
-        if self._closed:
-            return
-        self._closed = True
-        quit_now = getattr(self._response, "quit_now", None)
-        if quit_now is not None:
-            quit_now.set()
-        stream_task = getattr(self._response, "astream_task", None)
-        if stream_task is not None and not stream_task.done():
-            stream_task.cancel()
-            with suppress(asyncio.CancelledError):
-                await stream_task
-            return
-        await self._response.aclose()
-
-
-class ClaudeCurlClient:
-    def __init__(
-        self,
-        *,
-        session: Any | None = None,
-        connect_timeout_seconds: float = 30.0,
-        idle_timeout_seconds: float = 300.0,
-    ) -> None:
-        self._session = session or AsyncSession(
-            max_clients=50,
-            allow_redirects=False,
-        )
-        self._timeout = (connect_timeout_seconds, idle_timeout_seconds)
-
-    def build_request(
-        self,
-        method: str,
-        url: str,
-        *,
-        params: Sequence[tuple[str, str]] = (),
-        headers: Mapping[str, str] | None = None,
-        content: bytes = b"",
-    ) -> ClaudeCurlRequest:
-        return ClaudeCurlRequest(
-            method=method,
-            url=url,
-            params=tuple(params),
-            headers=dict(headers or {}),
-            content=content,
-        )
-
-    async def send(
-        self,
-        request: ClaudeCurlRequest,
-        *,
-        stream: bool = True,
-    ) -> ClaudeCurlResponse:
-        headers = dict(request.headers)
-        for key in tuple(headers):
-            if key.casefold() == "accept-encoding":
-                headers.pop(key)
-        headers["accept-encoding"] = "identity"
-        try:
-            response = await self._session.request(
-                method=request.method,
-                url=request.url,
-                params=list(request.params),
-                headers=headers,
-                data=request.content,
-                stream=stream,
-                timeout=self._timeout,
-                allow_redirects=False,
-                accept_encoding="identity",
-            )
-        except RequestException as exc:
-            raise httpx.TransportError(str(exc)) from exc
-        return ClaudeCurlResponse(response)
-
-    async def aclose(self) -> None:
-        await self._session.close()
+ClaudeCurlRequest = CurlRequest
+ClaudeCurlResponse = CurlResponse
+ClaudeCurlClient = CurlClient
